Issue-205: Moved extension package to core module.
