+++ app/build.gradle
@@ -75,9 +75,6 @@ dependencies {
     implementation libraries.gson
     implementation libraries.retrofitCoroutines
 
-    implementation libraries.glide
-    annotationProcessor libraries.glideCompiler
-
     implementation libraries.timber
     implementation libraries.stetho
     implementation libraries.stethoOkHttp
