--- app/src/main/java/com/kirchhoff/movies/extensions/Fragment.kt
@@ -1,6 +1,0 @@
-package com.kirchhoff.movies.extensions
-
-import androidx.annotation.DimenRes
-import androidx.fragment.app.Fragment
-
-fun Fragment.getSizeFromRes(@DimenRes size: Int) = resources.getDimensionPixelSize(size)
