+++ app/src/main/java/com/kirchhoff/movies/ui/screens/details/movie/adapters/TrailerVH.kt
@@ -3,9 +3,9 @@ package com.kirchhoff.movies.ui.screens.details.movie.adapters
 import android.view.ViewGroup
 import android.widget.ImageView
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.extensions.downloadYoutubePoster
+import com.kirchhoff.movies.core.extensions.inflate
 import com.kirchhoff.movies.data.ui.details.movie.UITrailer
-import com.kirchhoff.movies.extensions.downloadYoutubePoster
-import com.kirchhoff.movies.extensions.inflate
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class TrailerVH(parent: ViewGroup) : BaseVH<UITrailer>(parent.inflate(R.layout.item_trailer)) {
