+++ app/src/main/java/com/kirchhoff/movies/ui/screens/main/persons/adapter/PersonVH.kt
@@ -3,9 +3,9 @@ package com.kirchhoff.movies.ui.screens.main.persons.adapter
 import android.view.ViewGroup
 import android.widget.ImageView
 import com.kirchhoff.movies.R
+import com.kirchhoff.movies.core.extensions.downloadAvatar
+import com.kirchhoff.movies.core.extensions.inflate
 import com.kirchhoff.movies.data.ui.main.UIPerson
-import com.kirchhoff.movies.extensions.downloadAvatar
-import com.kirchhoff.movies.extensions.inflate
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class PersonVH(parent: ViewGroup) : BaseVH<UIPerson>(parent.inflate(R.layout.item_person)) {
