+++ app/src/main/java/com/kirchhoff/movies/ui/views/keywords/adapter/KeywordVH.kt
@@ -3,7 +3,7 @@ package com.kirchhoff.movies.ui.views.keywords.adapter
 import android.view.ViewGroup
 import android.widget.TextView
 import com.kirchhoff.movies.R
-import com.kirchhoff.movies.extensions.inflate
+import com.kirchhoff.movies.core.extensions.inflate
 import com.kirchhoff.movies.ui.utils.recyclerView.BaseVH
 
 class KeywordVH(parent: ViewGroup) : BaseVH<String>(parent.inflate(R.layout.item_keyword)) {
