+++ core/src/main/java/com/kirchhoff/movies/core/extensions/ViewGroup.kt
@@ -1,0 +1,7 @@
+package com.kirchhoff.movies.core.extensions
+
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+
+fun ViewGroup.inflate(layoutRes: Int): View = LayoutInflater.from(context).inflate(layoutRes, this, false)
